+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Batch change for payment and adress info.
+
 ## [3.1.1] - 2020-07-08
 
 ### Removed